Trait rust_server::database::database_base::Database

pub trait Database: Send + Sync {
    // Required methods
    fn create_user<'life0, 'async_trait>(
        &'life0 self,
        user: RegisterUser
    ) -> Pin<Box<dyn Future<Output = Result<UserCreated, SignupError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_admin<'life0, 'async_trait>(
        &'life0 self,
        user: RegisterUser
    ) -> Pin<Box<dyn Future<Output = Result<UserCreated, SignupError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn login<'life0, 'async_trait>(
        &'life0 self,
        user: LoginUser
    ) -> Pin<Box<dyn Future<Output = Result<UserLoggedIn, AuthenticationError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_user_by_id<'life0, 'async_trait>(
        &'life0 self,
        id: usize
    ) -> Pin<Box<dyn Future<Output = Option<User>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn add_accessible_article_to_id<'life0, 'async_trait>(
        &'life0 self,
        id: usize,
        article: PaywallArticle
    ) -> Pin<Box<dyn Future<Output = Result<(), ()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn user_id_has_access_by_link<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: usize,
        link: &'life1 str
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn user_id_is_verified<'life0, 'async_trait>(
        &'life0 self,
        id: usize
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn confirm_email_for_user_id<'life0, 'async_trait>(
        &'life0 self,
        id: usize
    ) -> Pin<Box<dyn Future<Output = Result<(), ()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn delete_user_by_id<'life0, 'async_trait>(
        &'life0 self,
        id: usize
    ) -> Pin<Box<dyn Future<Output = Result<(), ()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_paywall_articles_for_user_id<'life0, 'async_trait>(
        &'life0 self,
        id: usize
    ) -> Pin<Box<dyn Future<Output = Option<Vec<PaywallArticle>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn create_user<'life0, 'async_trait>( &'life0 self, user: RegisterUser ) -> Pin<Box<dyn Future<Output = Result<UserCreated, SignupError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create new user and directly provide login JWT after successful signup

use tokio::runtime::Runtime;
use rust_server::database::{Database, InMemoryDb};
use rust_server::models::RegisterUser;
use rust_server::security::NonHashing;

let db: InMemoryDb<NonHashing> = InMemoryDb::new("123".to_string());
let new_user = RegisterUser {
     email: "test@test.com".to_string(),
     username: "test".to_string(),
     password: "insecure password".to_string(),
     password_repeat: "insecure password".to_string()
     };

let rt = Runtime::new().unwrap();
let registered = rt.block_on(db.create_user(new_user)).unwrap();
assert_eq!(registered.user_id, 0);
assert_eq!(registered.email, "test@test.com");
assert_eq!(registered.username, "test");

fn create_admin<'life0, 'async_trait>( &'life0 self, user: RegisterUser ) -> Pin<Box<dyn Future<Output = Result<UserCreated, SignupError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create user with admin privileges

fn login<'life0, 'async_trait>( &'life0 self, user: LoginUser ) -> Pin<Box<dyn Future<Output = Result<UserLoggedIn, AuthenticationError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Match provided credentials against stored credentials and grant access token if they are matching.

use tokio::runtime::Runtime;

use rust_server::models::{RegisterUser, LoginUser};
use rust_server::database::{Database, InMemoryDb};
use rust_server::security::NonHashing;

let db: InMemoryDb<NonHashing> = InMemoryDb::new("123".to_string());
let new_user = RegisterUser {
     email: "test@test.com".to_string(),
     username: "test".to_string(),
     password: "insecure password".to_string(),
     password_repeat: "insecure password".to_string()
};

let rt = Runtime::new().unwrap();
let _ = rt.block_on(db.create_user(new_user));

let login_user = LoginUser {
     email: "test@test.com".to_string(),
     password: "insecure password".to_string()
};
let logged_in = rt.block_on(db.login(login_user)).unwrap();
assert_eq!(logged_in.username, "test");

fn get_user_by_id<'life0, 'async_trait>( &'life0 self, id: usize ) -> Pin<Box<dyn Future<Output = Option<User>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Provide a user_id and get the corresponding User object if it exists. If not, returns None.

Primarily used for matching a JWT user id to the respective user

use tokio::runtime::Runtime;

use rust_server::models::{RegisterUser, LoginUser};
use rust_server::database::{Database, InMemoryDb};
use rust_server::security::NonHashing;

let db: InMemoryDb<NonHashing> = InMemoryDb::new("123".to_string());
let new_user = RegisterUser {
     email: "test@test.com".to_string(),
     username: "test".to_string(),
     password: "insecure password".to_string(),
     password_repeat: "insecure password".to_string()
};

let rt = Runtime::new().unwrap();
let _ = rt.block_on(db.create_user(new_user));

let user_option = rt.block_on(db.get_user_by_id(1));
assert!(user_option.is_none());

let user = rt.block_on(db.get_user_by_id(0)).unwrap();
assert_eq!(user.username, "test");

fn add_accessible_article_to_id<'life0, 'async_trait>( &'life0 self, id: usize, article: PaywallArticle ) -> Pin<Box<dyn Future<Output = Result<(), ()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Add a PaywallArticle to the list of accessible_articles for a user with a given user_id.

use tokio::runtime::Runtime;

use rust_server::models::{RegisterUser, LoginUser, PaywallArticle};
use rust_server::price::Price;
use rust_server::database::{Database, InMemoryDb};
use rust_server::security::NonHashing;

let db: InMemoryDb<NonHashing> = InMemoryDb::new("123".to_string());
let new_user = RegisterUser {
     email: "test@test.com".to_string(),
     username: "test".to_string(),
     password: "insecure password".to_string(),
     password_repeat: "insecure password".to_string()
};

let rt = Runtime::new().unwrap();
let _ = rt.block_on(db.create_user(new_user));

let article = PaywallArticle::new(
     "test identifier".to_string(),
     "/test/test-article".to_string(),
     "test title".to_string(),
     Price::from_currency_string(10, "USD").unwrap()
);

let add_result = rt.block_on(db.add_accessible_article_to_id(0, article));

assert!(add_result.is_ok());

fn user_id_has_access_by_link<'life0, 'life1, 'async_trait>( &'life0 self, id: usize, link: &'life1 str ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

For a given user_id, check if there exists a PaywallArticle that contains the provided link; in the future, will probably use identifier to check access.

use tokio::runtime::Runtime;

use rust_server::models::{RegisterUser, LoginUser, PaywallArticle};
use rust_server::price::Price;
use rust_server::database::{Database, InMemoryDb};
use rust_server::security::NonHashing;

let db: InMemoryDb<NonHashing> = InMemoryDb::new("123".to_string());
let new_user = RegisterUser {
     email: "test@test.com".to_string(),
     username: "test".to_string(),
     password: "insecure password".to_string(),
     password_repeat: "insecure password".to_string()
};

let rt = Runtime::new().unwrap();
let _ = rt.block_on(db.create_user(new_user));

let article = PaywallArticle::new(
     "test identifier".to_string(),
     "/test/test-article".to_string(),
     "test title".to_string(),
     Price::from_currency_string(10, "USD").unwrap()
);

let _ = rt.block_on(db.add_accessible_article_to_id(0, article));

let has_access = rt.block_on(db.user_id_has_access_by_link(0, "/test/test-article"));

assert!(has_access);

fn user_id_is_verified<'life0, 'async_trait>( &'life0 self, id: usize ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Check if user has been verified (currently via mail; in the future probably also via Google/Github Login)

use tokio::runtime::Runtime;

use rust_server::models::RegisterUser;
use rust_server::database::{Database, InMemoryDb};
use rust_server::security::NonHashing;

let db: InMemoryDb<NonHashing> = InMemoryDb::new("123".to_string());
let new_user = RegisterUser {
     email: "test@test.com".to_string(),
     username: "test".to_string(),
     password: "insecure password".to_string(),
     password_repeat: "insecure password".to_string()
};

let rt = Runtime::new().unwrap();
let _ = rt.block_on(db.create_user(new_user));

let is_verified = rt.block_on(db.user_id_is_verified(0));
assert!(!is_verified);

fn confirm_email_for_user_id<'life0, 'async_trait>( &'life0 self, id: usize ) -> Pin<Box<dyn Future<Output = Result<(), ()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Verifies the user

use tokio::runtime::Runtime;

use rust_server::models::RegisterUser;
use rust_server::database::{Database, InMemoryDb};
use rust_server::security::NonHashing;

let db: InMemoryDb<NonHashing> = InMemoryDb::new("123".to_string());
let new_user = RegisterUser {
     email: "test@test.com".to_string(),
     username: "test".to_string(),
     password: "insecure password".to_string(),
     password_repeat: "insecure password".to_string()
};

let rt = Runtime::new().unwrap();
let _ = rt.block_on(db.create_user(new_user));

let _ = rt.block_on(db.confirm_email_for_user_id(0));

let is_verified = rt.block_on(db.user_id_is_verified(0));
assert!(is_verified);

fn delete_user_by_id<'life0, 'async_trait>( &'life0 self, id: usize ) -> Pin<Box<dyn Future<Output = Result<(), ()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Delete a user by providing their user_id.

use tokio::runtime::Runtime;

use rust_server::models::RegisterUser;
use rust_server::database::{Database, InMemoryDb};
use rust_server::security::NonHashing;

let db: InMemoryDb<NonHashing> = InMemoryDb::new("123".to_string());
let new_user = RegisterUser {
     email: "test@test.com".to_string(),
     username: "test".to_string(),
     password: "insecure password".to_string(),
     password_repeat: "insecure password".to_string()
};

let rt = Runtime::new().unwrap();
let _ = rt.block_on(db.create_user(new_user));

let user_option = rt.block_on(db.get_user_by_id(0));
assert!(user_option.is_some());

let _ = rt.block_on(db.delete_user_by_id(0));
let user_option = rt.block_on(db.get_user_by_id(0));
assert!(user_option.is_none());

fn get_paywall_articles_for_user_id<'life0, 'async_trait>( &'life0 self, id: usize ) -> Pin<Box<dyn Future<Output = Option<Vec<PaywallArticle>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

If user_id exists, returns all PaywallArticles that the corresponding user has access to. If user doesn’t exist, returns None

use tokio::runtime::Runtime;

use rust_server::models::{RegisterUser, LoginUser, PaywallArticle};
use rust_server::price::Price;
use rust_server::database::{Database, InMemoryDb};
use rust_server::security::NonHashing;

let db: InMemoryDb<NonHashing> = InMemoryDb::new("123".to_string());
let new_user = RegisterUser {
     email: "test@test.com".to_string(),
     username: "test".to_string(),
     password: "insecure password".to_string(),
     password_repeat: "insecure password".to_string()
};

let rt = Runtime::new().unwrap();
let _ = rt.block_on(db.create_user(new_user));

let articles = rt.block_on(db.get_paywall_articles_for_user_id(0)).unwrap();
assert_eq!(articles.len(), 0);

let article = PaywallArticle::new(
     "test identifier".to_string(),
     "/test/test-article".to_string(),
     "test title".to_string(),
     Price::from_currency_string(10, "USD").unwrap()
);

let _ = rt.block_on(db.add_accessible_article_to_id(0, article));
let articles = rt.block_on(db.get_paywall_articles_for_user_id(0)).unwrap();
assert_eq!(articles.len(), 1);