Function rust_server::handlers::ai::eventstorming::testable_handle_eventstorming_post

pub fn testable_handle_eventstorming_post(
    config: Config,
    payload: AIEventStormingRequest
) -> (StatusCode, Json<AIEventStormingResponse>)

Handles the AI event storming request.

This function runs the event storming script with the given prompt and sends an response back to the client.

Arguments

config - The server Config.

payload - The AIEventStormingRequest payload.

Returns

A tuple containing the [StatusCode] and the AIEventStormingResponse as JSON. The successfull response contains a message Event Storming successfull..

Debug

This function logs the following:

An info message when the event storming script is trying to run as info in the format Trying to run event storming script.

An info message when the event storming script is successfull as info in the format Event Storming successfull.

An error message when the event storming script fails as error in the format Error: <error>.