Module rust_server::paywall

This module holds the key logic for a pay-per-article paywall In here, things have become quite abstract but with the (hopeful) advantage of being fairly extensible for the future.

Some ideas including paywalling other file-formats, quarto books and possibly also a way to implement a subscription based paywall

In general, this module consists of 3 modules:

A recursive filereader: Reads through a complete folder and loads filepaths and filenames for all given filetypes

A file manipulator: Manipulates the files by user provided functions and allows to assign the output of each function to a session based property. E.g. we read all HTML quarto outputs, remove paywalled content and add respective notifications to the document, based on the user privileges (logged out, logged in, paid for article).

A paywall extraction function: Extracts the properties of the paywalled item (identifier, price, link) from the object that was loaded from the filereader. Currently, this is just done for html/text but we could potentially also implement that for videos or pdfs

Re-exports

pub use paywall_server::*;

pub use paywall_server_factory::*;

pub use quarto_paywall::*;

pub use recursive_filereader::*;

pub use session_conditional_manipulation::*;

pub use session_conditional_object::*;

Modules

paywall_server

Main abstraction of the paywall module. PaywallServer is a trait that contains all the necessary functions to handle paywalled content in an abstract way. Right now, an std::collections::HashMap is the implementation of choice but other, more efficient data structures could be implemented as well in the future.

paywall_server_factory

Factory method to create a paywall server from its key components (i.e. recursive filereader, object manipulator, paywall extractor)

quarto_paywall

Unifying factory method to create a paywall for quarto HTML pages; this is not necessarily the only use case and other types of content, e.g. documents or videos could also be put behind a paywall in the future. Only a few changes and/or additions should be necessary to accomplish the latter.

recursive_filereader

Simple filereader to recursively read and load all files from a designated directory (e.g. the quarto export directory)

session_conditional_manipulation

Abstractions to manipulate data and assign the output to a SessionConditionalObject, e.g. to serve variations of the same object based on different values of a user session (e.g. their privileges)

session_conditional_object

Stores variations of the same object and returns a variation based on given properties of the user session