Function rust_server::utils::postgres::find_answer

pub async fn find_answer(
    info: &DBConnectionInfo,
    prompt_id: i32
) -> Result<(i32, String)>

Utility function to easily find an answer in the database.

The function returns the answer with the highest rating and filters out any answer with a rating less than or equal to 0.

Arguments

info - The DBConnectionInfo struct containing the database connection details.

prompt_id - The ID of the prompt to which the answer belongs.

Returns

A Result containing a tuple with the ID of the answer and the answer itself or an DBError of the operation.

Errors

DBError::ConnectionError - If there is an error while connecting to the database.

DBError::ExtensionError - If there is an error while installing the vector extension.

DBError::TableCreationError - If there is an error while creating the table.

DBError::QueryError - If there is an error while trying to find the answer.