Trait rust_server::utils::AdvancedEditable

pub trait AdvancedEditable {
    // Required methods
    fn insert_at_index_or_push(
        &mut self,
        selector: &Selector,
        target: Node,
        index: usize
    ) -> &mut Self;
    fn insert_before_selector_or_push(
        &mut self,
        selector: &Selector,
        target: Node,
        before_selector: &Selector
    ) -> &mut Self;
}

Required Methods

fn insert_at_index_or_push( &mut self, selector: &Selector, target: Node, index: usize ) -> &mut Self

Insert node at index in the children all elements that match the selector. If index lies outside the range of the children, target is inserted as the last element of children.

use html_editor::{parse, Node};
use html_editor::operation::*;
use rust_server::utils::AdvancedEditable;

let html = r#"<div><ul><li>1</li><li>3</li></ul></div>"#;

let selector = Selector::from("ul");
let html = parse(html)
    .unwrap()
    .insert_at_index_or_push(&selector, Node::new_element(
        "li",
        vec![],
        vec![Node::Text("2".to_string())]
    ), 1)
    .html();
assert_eq!(html, r#"<div><ul><li>1</li><li>2</li><li>3</li></ul></div>"#)

fn insert_before_selector_or_push( &mut self, selector: &Selector, target: Node, before_selector: &Selector ) -> &mut Self

Insert node before (at the same hierarchical level) all (recursive) all child elements of selector that match the before_selector. If before_selector is not found, appends to all selector nodes.

use html_editor::{parse, Node};
use html_editor::operation::*;
use rust_server::utils::AdvancedEditable;

let html = r#"<div><ul><li>1</li><li class="third">3</li></ul></div>"#;

let selector = Selector::from("ul");
let before_selector = Selector::from("li.third");
let html = parse(html)
    .unwrap()
    .insert_before_selector_or_push(&selector, Node::new_element(
        "li",
        vec![],
        vec![Node::Text("2".to_string())]
    ), &before_selector)
    .html();
assert_eq!(html, r#"<div><ul><li>1</li><li>2</li><li class="third">3</li></ul></div>"#)