Crate rust_server

The Rust Axum AI Server for the Prooph-Board Chatbot Chaty.

This server is used to run the AI model for the chatbot and handle requests from the chatbot frontend.

The server is built using the Axum framework and uses the Tokio runtime for asynchronous tasks.

The server uses the PostgreSQL database for storing prompts and responses and uses Python Scripts to handle the AI model.

The server is tested using Rust’s built-in testing framework.

Routes

The routes are defined in the router module.

The server has the following routes:

ai/prompt - POST - Handles the AI prompt request.

ai/rating - POST - Handles the AI rating request.

ai/regenerate - POST - Handles the AI regenerate request.

ai/eventstorming - POST - Handles the AI Eventstorming request.

Configuration

The server configuration is defined in the config Struct.

Environment Variables

The following environment variables are used to load the configuration, please make a .env file in the root directory of the project and put at least the bold ones in it:

DB_HOST: The host/ip of the PostgreSQL database.

DB_NAME: The name of the PostgreSQL database.

DB_USER: The username for the PostgreSQL database.

DB_PASSWORD: The password for the PostgreSQL database.

LLM_SERVER: The connection url for the LLM Server (Ollama).

IP: The IP address of the server. Defaults to 127.0.0.1.

PORT: The port of the server. Defaults to 3000.

DB_PORT: The port of the PostgreSQL database. Defaults to 5432.

Database

The Server uses a PostgreSQL database with the vector extension. There is a docker container available for the database called pgvector which can be found at: https://hub.docker.com/r/pgvector/pgvector. That container needs the environment variable POSTGRES_PASSWORD to be set.

Modules

config

This module contains the configuration settings for the server.

handlers

This module contains the request handlers, handling the requests from the chatbot frontend coming via HTTP to the Server.

models

This module contains the data models used by the server.

utils

This module contains utility functions used by the server like basic database functions and python script execution.